LIGHTS!
Finally!
So...
Each light has an on/off state.
This state can be changed individually.
All the lights can be turned off by the fusebox. If the fusebox is missing the fuse, none of the lights can be turned on. If the fusebox recieves the fuse, all lights can be turned on alltogether or individually :)
This took so long wtf
+ Need to link up lights so that whole rooms can be lit up by a switch